+++ src/Common/Utilities/src/TextObjectFormatter.h
@@ -35,7 +35,7 @@ namespace BitFunnel
         void Format(bool value);
         void Format(int value);
         void Format(unsigned value);
-        void Format(unsigned long value);
+        void Format(size_t value);
         void Format(double value);
         void Format(char const * value);
         void FormatStringLiteral(char const * value);
